Created landing pages for adwords
I created two landing pages for our adwords campains, one for archery
and another for fly fishing.  I chose both of these because they had
the highest search volume.

I also created two important partials: first
a 'trifecta' partial that was origionlay just at the bottom of the home
page, trifecta because it covers Parents, Campers, and Staff.  It is located
in the new shared folder, for partials used across directories.  I also
created a partial for use on the bottom half of landing pages that
covers both "why a tradtional camp" and "why WLC".  This allows
the quick deployment of additional landing pages as needed, and they all
will contain the inmportant info about camp.