Created thank you page
I created a thank you page for display after someone finishes filling
out the contact us form.  It is hooked up to adwords and registers
it as a lead.

Adwords conversion on contact page

I added a script to the contact page that allows adwords to tacks
that as a "view of critical page" in adwords.

Removed contact us adwords script.